+++ php/class-asset-manager-scripts.php
@@ -209,6 +209,10 @@ class Asset_Manager_Scripts extends Asset_Manager {
 	 * @return array
 	 */
 	public function pre_add_asset( $script ) {
+		// Register/enqueue script with new 'strategy' argument if load method is async or defer.
+		if ( in_array( $script['load_method'], $this->wp_enqueue_methods, true ) ) {
+			$script['args']['strategy'] = $script['load_method'];
+		}
 		return $script;
 	}
 
@@ -253,7 +257,7 @@ class Asset_Manager_Scripts extends Asset_Manager {
 	 */
 	public function add_to_async( $script ) {
 		if (
-			( 'defer' === $script['load_method'] || 'async' === $script['load_method'] || 'async-defer' === $script['load_method'] ) &&
+			( 'async-defer' === $script['load_method'] ) &&
 			! in_array( $script['handle'], $this->async_scripts, true )
 		) {
 			$this->async_scripts[] = $script['handle'];