+++ php/class-asset-manager.php
@@ -299,13 +299,24 @@ abstract class Asset_Manager {
 			// Enqueue asset if applicable.
 			if ( in_array( $args['load_method'], $this->wp_enqueue_methods, true ) && empty( $args['loaded'] ) ) {
 				if ( function_exists( $wp_enqueue_function ) ) {
+					$enqueue_options = [];
+					if ( 'style' === $args['type'] ) {
+						$enqueue_options = $args['media'];
+					} else {
+						$enqueue_options['in_footer'] = $args['in_footer'];
+						if ( in_array( $args['load_method'], [ 'async', 'defer' ], true ) ) {
+							$enqueue_options['strategy'] = $args['load_method'];
+						}
+					}
+
 					$wp_enqueue_function(
 						$args['handle'],
 						$args['src'],
 						$args['deps'],
 						$args['version'],
-						'style' === $args['type'] ? $args['media'] : $args['in_footer']
+						$enqueue_options
 					);
+
 					$args['loaded'] = true;
 				} else {
 					echo wp_kses_post( $this->format_error( $this->generate_asset_error( 'invalid_enqueue_function', false, $wp_enqueue_function ) ) );